# Use all unsolved pieces on the offending line as expand pieces.